Change the scopeName of javascriptreact to "source.js"
Because of 74258ca the scopeName `source.js.jsx` doesn't exist anymore so to make #30 work we need to map the `javascriptreact` grammar to `source.js`.

Also, added the extension `.js` to the language identifier so it can work with js files that have `Javascript React` as the language id. too.